+++ CHANGELOG.md
@@ -10,12 +10,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [D] Changes is `Cere Dev` Runtime
 
 ## [vNext]
-- [C,D] `pallet-ddc-customers`: implemented bucket removal
-- Added ChargeError event to payout pallet
 
 ### Added
 
-- ...
+- Missing storage migrations to Staking pallet
 
 ### Changed